+++ src/modules/alert.js
@@ -5,7 +5,7 @@ module.exports = ({ bot, knex, config, commands }) => {
       await thread.postSystemMessage("Cancelled new message alert");
     } else {
       await thread.addAlert(msg.author.id);
-      await thread.postSystemMessage(`Pinging ${msg.author.username} when this thread gets a new reply`);
+      await thread.postSystemMessage(`Pinging ${msg.author.nick || config.useDisplaynames ? msg.author.globalName || msg.author.username : msg.author.username} when this thread gets a new reply`);
     }
   }, { allowSuspended: true });
 };