+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 For instructions on how to update the bot, see **[✨ Updating the bot](docs/updating.md)**
 
+## v3.8.0 (2024-MM-DD)
+* Updated eris to 0.18.0
+* Add new config option `useDisplaynames` to use a users display name instead of their username
+  * Defaults to `on`
+  * `useNicknames` will override this in most cases
+
 ## v3.7.0 (2023-09-12)
 * Added support for Node.js 18 and higher
 * Added `!realreply` / `!rr` command ([#763](https://github.com/Dragory/modmailbot/pull/763))